zadanie1/requirements.txt
pydantic
miskibin
redis
geocoder
pandas
pyaml